npme docs edit 5
edited platforms docs. PLEASE NOTE: some of the language here is confusing.
in the azure (line 10) and vagrant (line 11) docs, the phrase "...and then we'll need to" is used in the middle of a numbered list of steps. Does this mean npm needs to do something? if so, should we be asking the reader to contact npm? or is this something they should be doing?